+++ docs/misc/changelog.md
@@ -11,6 +11,8 @@
 
 ### New Features:
 
+- Added `squash_mean_actions` policy option for A2C/PPO to tanh-squash the mean of `DiagGaussianDistribution`
+
 ### Bug Fixes:
 - Fixed deprecated error Taxi-v3 from gymnasium v1.3.0 in tests